Show_game_result

+++ VolleyManagement.Domain/GamesAggregate/Game.cs
@@ -40,6 +40,6 @@
         /// <summary>
         /// Gets or sets the round of the game in the tournament.
         /// </summary>
-        public int Round { get; set; }
+        public byte Round { get; set; }
     }
 }
